Change in "Snippet" methods
Added more restricts in the type of variables passed through the
methods.
"Threshold" now accepts integers instead of doubles, it also verifies
that the value is positive.
"Gamma" verifies that the value is positive.
"Contrast" accepts integers instead of doubles and verifies that both
the values are positive as well.