MAJOR UPDATE #1
-"Snippet" class is now in a new package called "Snippet".
-A new private class "Signature" is added in "Snippet" class, its
constructor is private, an object is needed to be passed to every
function in every interface implemented so far ( and every interface to
be made in the future).
-All classes that implements an interface had the extra parameter added
to them.
-Inside the body of each function in each class, the hashcode of the
passed "Signature" object is called to throw a "NullException" if the
passed object was null and not the valid one, which can only be called
by using "Snippet" class methods.
-This simulates the "friend" meachanism in C++.